refactor: move tests to the zmessaging module
This commit also removes all old ignored tests, as well as the
integration, mocked and unit modules. Many tests were either fixed or
removed depending on their state.

The motivation for this was a simplifed build setup and it also allows
us to run tests inside IntelliJ.

For best results, also set the project SDK to Java 1.8 and not the
Android SDK.